requirements.txt
aio-pika==0.10.1
asyncio==3.4.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pika==0.10.0
pillow==11.0.0
propcache==0.4.1
psutil==6.1.0
py-cpuinfo==9.0.0
pygame==2.6.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
shortuuid==1.0.13
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.6
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.26
ultralytics-thop==2.0.10
urllib3==2.2.3
websockets==13.1
yarl==1.22.0
